Crear Controller, Services i lògica d'afegir Exercici, treure exercic…

+++ src/main/java/com/mygym/models/Serie.java
@@ -2,10 +2,9 @@ package com.mygym.models;
 
 import com.mygym.security.Auditable;
 
-//Classe Feble d'Exercici
 public class Serie extends Auditable {
     private int repeticions;
-    private int pes;
+    private double pes;
     private int descans;
 
     // Getters i setters
